change dto

+++ src/main/java/com/example/msinternauticacrm/models/Boat.java
@@ -16,7 +16,7 @@ public class Boat {
     @Id
     private String id;
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idBoat;
+    private int idBoat;
 
     private List<String> photos;
